requirements.txt
torch==1.6.0
torchvision==0.7.0
scipy==1.5.2
jsonpatch==1.26
visdom==0.1.8.9
torchtoolbox==0.1.5
matplotlib==3.3.2
imutils==0.5.3
scikit-image==0.17.2
--extra-index-url https://developer.download.nvidia.com/compute/redist/ nvidia-dali-cuda100
opencv-python-headless==4.4.0.44
cmake==3.18.2
dlib==19.19.0
onnxruntime==1.5.1
docker==4.3.1
pandas==1.1.3
screeninfo==0.6.5
future==0.18.2
numpy==1.19.2
pillow==8.0.0
jsonpointer==2.0
requests==2.24.0
tornado==6.0.4
pyzmq==19.0.2
six==1.15.0
torchfile==0.1.0
websocket-client==0.57.0
lmdb==1.0.0
pyarrow==1.0.1
tqdm==4.50.2
scikit_learn==0.23.2
kiwisolver==1.2.0
certifi==2020.06.20
cycler==0.10.0
python-dateutil==2.8.1
pyparsing==2.4.7
tifffile==2020.10.1
PyWavelets==1.1.1
imageio==2.9.0
networkx==2.5
protobuf==3.13.0
pytz==2020.1
dataclasses==0.6
urllib3==1.25.10
idna==2.10
chardet==3.0.4
joblib==0.17.0
threadpoolctl==2.1.0
decorator==4.4.2